Add a POC cumbersome fix for Moodle3_9
This version works when tested manually on all
versions of Moodle and PHP, but more important
on Travis.

It fixes that CI test fail that have been detected on
Moodle3_9 only; while it works perfectly, when tested
manually.

This is a very strange behavior that has been detected
when running behat tests. The MathType client-side
render does not work, ( nor Mathjax ), because the
filter fails to convert right from SafeMathML to
standard MathML.

It generates a wrong MathML format so the WIRISPlugins.js
is unable to render it right. It shows an empty image.

This only happens when running on behat scenario, but works
fine if it's tested manually. In moodle3_9 works too. :shrug:

Also:

- Fix MathJax tests to be compatible with these changes.

- Fix lint errors on shome gherkin  tests.